add quant_nn.py for the modularity. test=develop
