+++ python/setup.py.in
@@ -160,6 +160,7 @@ packages=['paddle',
           'paddle.fluid.contrib.slim.graph',
           'paddle.fluid.contrib.slim.prune',
           'paddle.fluid.contrib.slim.quantization',
+          'paddle.fluid.contrib.slim.quantization.dyquant',
           'paddle.fluid.contrib.slim.distillation',
           'paddle.fluid.contrib.slim.nas',
           'paddle.fluid.contrib.slim.searcher',
